+++ Common/Source/GrowlDefinesInternal.m
@@ -1,9 +1,0 @@
-//
-//  GrowlDefines.m
-//  Growl
-//
-//  Created by Peter Hosey on 2010-07-10.
-//  Copyright 2010 The Growl Project. All rights reserved.
-//
-
-NSString *const GrowlErrorDomain = @"GrowlErrorDomain";