+++ Plugins/Actions/Prowl/Prowl/PRAPIKey.h
@@ -1,9 +1,0 @@
-#import <Foundation/Foundation.h>
-
-@interface PRAPIKey : NSObject <NSCoding>
-
-@property (nonatomic, assign) BOOL enabled;
-@property (nonatomic, assign) BOOL validated;
-@property (nonatomic, copy) NSString *apiKey;
-
-@end