+++ Plugins/Actions/Prowl/Prowl/PRAction.h
@@ -1,7 +1,0 @@
-#import <GrowlPlugins/GrowlActionPlugin.h>
-#import "PRDefines.h"
-#import "PRAPIKey.h"
-
-@interface PRAction : GrowlActionPlugin <GrowlDispatchNotificationProtocol, GrowlUpgradePluginPrefsProtocol>
-
-@end